code

+++ 0x00-hello_world/6-size.c
@@ -1,4 +1,11 @@
 #include <stdio.h>
+/**
+ * main - print the string in the put  function
+ *
+ * Description using the main function
+ * this program prints "programming is like multilingual puzzle
+ * Return 0
+ */
 int main(void)
 {
 int a;
@@ -6,10 +13,10 @@ log int b
 long long  int c;
 char d;
 float f;
-printf("size of a char:%1u byte(s)\n", (unsigned long)sizeof(d));
-printf("size of an int:%1u byte(s)\n", (unsigned long)sizeof(a));
-printf("size of a long int:%1u byte(s)\n", (unsigned long)sizeof(b));
-printf("size of a long long int:%1u byte(s)\n", (unsigned long)sizeof(c));
-printf("size of a float:%1u byte(s)\n", (unsigned long)sizeof(f));
+printf("size of a char:%1d byte(s)\n", sizeof(d));
+printf("size of an int:%1d byte(s)\n", sizeof(a));
+printf("size of a long int:%1d byte(s)\n", sizeof(b));
+printf("size of a long long int:%1d byte(s)\n", sizeof(c));
+printf("size of a float:%1d byte(s)\n", sizeof(f));
 return (0);
 }
